Turn off default features of ahash as they're not needed
The default feature set of the ahash crate is ["std"][1]. The std
feature enables features which require the standard library, namely
`AHashMap` and `AHashSet`. DataFusion currently only uses `AHasher`,
`CallHasher`, and `RandomState`, none of which require the standard
library.

This gives more control to projects depending on datafusion to minimize
the amount of code they depend on.

[1]: https://github.com/tkaitchuck/aHash/blob/e77cab8c1e15bfc9f54dfd28bd8820c2a7bb27c4/Cargo.toml#L24-L25