Benchmark query cancellation
Connects to #14036.

This benchmark loads multiple files into an in-memory object store,
starts a datafusion query in a new tokio runtime, lets the query run for
an amount of time, cancels the query, and measures how long it takes to
drop the tokio runtime.

This demonstrates datafusion is likely not yielding often enough to
allow for timely query cancellation and freeing up of all resources.